requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
comm==0.1.3
contourpy==1.0.7
crashtest==0.4.1
crispy-bootstrap5==0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
django-allauth==0.52.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dulwich==0.21.3
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
fire==0.5.0
fonttools==4.39.3
fqdn==1.5.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.13.1
kiwisolver==1.4.4
lockfile==0.12.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.1.0
msgpack==1.0.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.3
oauthlib==3.2.2
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.6.2
pkginfo==1.9.6
platformdirs==3.6.0
pluggy==1.0.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rapidfuzz==2.14.0
rcssmin==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rjsmin==1.2.1
scipy==1.10.1
SecretStorage==3.3.3
Send2Trash==1.8.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
termcolor==2.2.0
terminado==0.17.1
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
tomlkit==0.11.7
tornado==6.2
traitlets==5.9.0
trove-classifiers==2023.3.9
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
whitenoise==6.3.0
widgetsnbextension==4.0.7
zipp==3.15.0
